GS: Fix a couple of 64 bit warnings

+++ pcsx2/GS/GS.cpp
@@ -1174,7 +1174,7 @@ void* fifo_alloc(size_t size, size_t repeat)
 
 	if (repeat >= countof(s_Next))
 	{
-		fprintf(stderr, "Memory mapping overflow (%zu >= %u)\n", repeat, countof(s_Next));
+		fprintf(stderr, "Memory mapping overflow (%zu >= %u)\n", repeat, static_cast<unsigned>(countof(s_Next)));
 		return vmalloc(size * repeat, false); // Fallback to default vmalloc
 	}
